requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
idna==2.6
pluggy==0.6.0
py==1.5.2
pytest==3.3.2
python-dateutil==2.6.1
python-telegram-bot==9.0.0
requests==2.18.4
requests-cache==0.4.13
schedule==0.5.0
six==1.11.0
urllib3==1.22
